feat: mejorar registro de solicitud con empleado y simplificar lógica

+++ src/main/java/ec/edu/ups/poo/models/interfaces/Calculable.java
@@ -1,5 +1,5 @@
 package ec.edu.ups.poo.models.interfaces;
 
 public interface Calculable {
-    int calcularCostoIVA();
+    double calcularCostoIVA();
 }
